[refactor] 불필요한 클래스 삭제
- 게시글 삭제 Application 삭제 -> PostService에 통합함
- 동일한 역할을 수행하는 Service가 2개나 존재하여 1개를 삭제하기로 판단함.
- 삭제가 진행되는 동안 데이터에 변함이 존재해서 안된다고 판단하여 @Transactional 어노테이션 추가

+++ src/main/java/org/board/exercise_board/post/service/PostService.java
@@ -1,6 +1,7 @@
 package org.board.exercise_board.post.service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.board.exercise_board.liked.service.FindByType;
@@ -65,8 +66,17 @@ public class PostService implements FindByType<Post> {
         .orElseThrow(() -> new PostCustomException(PostErrorCode.POST_IS_NOT_EXIST));
   }
 
-  public String deletePost(Post post) {
+  @Transactional
+  public String deletePost(String writerId, Long postId) {
+    Post post = this.find(postId);
+
+    // 로그인한 사용자가 작성자인지 확인
+    if(!Objects.equals(writerId, post.getUser().getLoginId())) {
+      throw new PostCustomException(PostErrorCode.NOT_HAVE_RIGHT);
+    }
+
     postRepository.delete(post);
+
     return "게시물 삭제가 완료되었습니다.";
   }
